refactor(task-microservice): Changed project folder structure (Golang Standards)
