+++ 1-task-microservice/go.mod
@@ -1,10 +1,14 @@
-module task-microservice
+module github.com/matheusdiass/taskmanager-microservices
 
 go 1.20
 
 require github.com/go-chi/chi/v5 v5.0.10
 
-require github.com/joho/godotenv v1.5.1
+require (
+	github.com/joho/godotenv v1.5.1
+	gorm.io/driver/postgres v1.5.3
+	gorm.io/gorm v1.25.5
+)
 
 require (
 	github.com/jackc/pgpassfile v1.0.0 // indirect
@@ -12,8 +16,6 @@ require (
 	github.com/jackc/pgx/v5 v5.4.3 // indirect
 	github.com/jinzhu/inflection v1.0.0 // indirect
 	github.com/jinzhu/now v1.1.5 // indirect
-	golang.org/x/crypto v0.9.0 // indirect
-	golang.org/x/text v0.9.0 // indirect
-	gorm.io/driver/postgres v1.5.3 // indirect
-	gorm.io/gorm v1.25.5 // indirect
+	golang.org/x/crypto v0.14.0 // indirect
+	golang.org/x/text v0.13.0 // indirect
 )
