+++ 1-task-microservice/internal/domain/contracts/create-task-repository.go
@@ -1,0 +1,7 @@
+package contracts
+
+import "github.com/matheusdiass/taskmanager-microservices/internal/domain/entity"
+
+type CreateTaskRepository interface {
+	Execute(task entity.Task) error
+}
