+++ 1-task-microservice/internal/domain/contracts/fetch-task-by-id-repository.go
@@ -1,0 +1,9 @@
+package contracts
+
+import (
+	"github.com/matheusdiass/taskmanager-microservices/internal/domain/dtos"
+)
+
+type FetchTaskByIdRepository interface {
+	Execute(id uint) (dtos.Task, error)
+}
