+++ 1-task-microservice/internal/domain/contracts/fetch-user-by-id-repository.go
