+++ 1-task-microservice/internal/domain/contracts/update-task-repository.go
@@ -1,0 +1,7 @@
+package contracts
+
+import "github.com/matheusdiass/taskmanager-microservices/internal/domain/entity"
+
+type UpdateTaskRepository interface {
+	Execute(id uint, task entity.Task) error
+}
