+++ 1-task-microservice/internal/domain/mocks/create-task-repository.go
@@ -1,0 +1,13 @@
+package mocks
+
+import "github.com/matheusdiass/taskmanager-microservices/internal/domain/entity"
+
+type createTaskRepository struct{}
+
+func NewCreateTaskRepository() createTaskRepository {
+	return createTaskRepository{}
+}
+
+func (r createTaskRepository) Execute(task entity.Task) error {
+	return nil
+}
