+++ 1-task-microservice/internal/domain/mocks/fetch-user-by-id-repository.go
