+++ 1-task-microservice/internal/main/contracts/http-adapter.go
@@ -1,0 +1,7 @@
+package contracts
+
+import "github.com/matheusdiass/taskmanager-microservices/internal/main/types"
+
+type HttpAdapter interface {
+	CreateRoute(route types.Route)
+}
