--- 1-task-microservice/src/domain/contracts/create-task-repository.go
@@ -1,7 +1,0 @@
-package contracts
-
-import "task-microservice/src/domain/entity"
-
-type CreateTaskRepository interface {
-	Execute(task entity.Task) error
-}
