--- 1-task-microservice/src/domain/contracts/fetch-task-by-id-repository.go
@@ -1,9 +1,0 @@
-package contracts
-
-import (
-	"task-microservice/src/domain/dtos"
-)
-
-type FetchTaskByIdRepository interface {
-	Execute(id uint) (dtos.Task, error)
-}
