--- 1-task-microservice/src/domain/contracts/update-task-repository.go
@@ -1,7 +1,0 @@
-package contracts
-
-import "task-microservice/src/domain/entity"
-
-type UpdateTaskRepository interface {
-	Execute(id uint, task entity.Task) error
-}
