--- 1-task-microservice/src/main/contracts/http-adapter.go
@@ -1,7 +1,0 @@
-package contracts
-
-import "task-microservice/src/main/types"
-
-type HttpAdapter interface {
-	CreateRoute(route types.Route)
-}
